// src/renderer/models/DeviceType.ts
export enum DeviceBrand {
    Bafang = 'bafang',
}

export enum DeviceInterface {
    UART = 'uart',
    CAN = 'can',
}

export enum DeviceType {
    Motor = 'motor',
    Display = 'display',
}

export enum DeviceName {
    BafangUartMotor = 'bafang_uart_motor',
    BafangUartDisplay = 'bafang_uart_display',
}